+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/entity/AdminMenu.java
@@ -6,12 +6,7 @@ import lombok.Data;
 import org.hibernate.annotations.Type;
 import org.hibernate.annotations.TypeDef;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 /**
  * Job admin menu and perms
@@ -35,7 +30,7 @@ public class AdminMenu {
      * Parent ID
      */
     @Column(name = "pid")
-    private Integer pid;
+    private Long pid;
 
     /**
      * Type. 1=menu 2=perm
